// client/src/models/terrains/terrain.types.ts
export enum TerrainType {
  GRASS = "grass",
  GRASS2 = "grass2",
  GRASS3 = "grass3",
  GRASS4 = "grass4",
  GRASS5 = "grass5",

  DESERT = "desert",
  DESERT2 = "desert2",
  DESERT3 = "desert3",
  DESERT4 = "desert4",
  DESERT5 = "desert5",

  SOIL = "soil",
  SOIL2 = "soil2",
  SOIL3 = "soil3",
  SOIL4 = "soil4",
  SOIL5 = "soil5",

  WAX = "wax",
  WAX2 = "wax2",
  WAX3 = "wax3",
  WAX4 = "wax4",
  WAX5 = "wax5",
}